migraciones corregidas

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->boolean('type');
-            $table->Integer('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->rememberToken();
             $table->timestamps();
